video: make it possible to scale/pan the video by arbitrary amounts
Add --video-pan-x, --video-pan-y, --video-scale options and properties.
See the additions to the manpage for description and semantics.

These transformations are intentionally done on top of panscan. Unlike
the (now removed) --panscanrange option, this doesn't affect the default
panscan behavior. (Although panscan itself becomes kind of useless if
the new options are used.)

Note: not sure about the behavior of the new options.